Fix SLL specialization join (float codes in Excel)
Codici nel foglio Excel erano float (101.0, 102.0) invece di interi.
Il zfill(4) non funzionava. Aggiunto cast astype(float).astype(int) prima
di zfill. Tutti 515 SLL ora hanno cod_classe/den_gruppo valorizzati.

+++ generate_sll_geojson.py
@@ -87,7 +87,7 @@ raw_spec.columns = ["cod_sll_spec","cod_sll_n","den_sll_spec",
                     "pop_spec","sup_spec"]
 raw_spec = raw_spec.dropna(subset=["cod_sll_spec"])
 raw_spec = raw_spec[raw_spec["cod_classe"].isin(["A","B","C","D"])]
-raw_spec["cod_sll_spec"] = raw_spec["cod_sll_spec"].astype(str).str.zfill(4)
+raw_spec["cod_sll_spec"] = raw_spec["cod_sll_spec"].astype(float).astype(int).astype(str).str.zfill(4)
 print(f"  {len(raw_spec)} SLL con specializzazione")
 
 # ── 4. Capoluogo per SLL dal JSON di composizione ──────────────────────────────
